006_auto_bot/001_code/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
schedule>=1.2.0
newspaper3k>=0.2.8
lxml[html_clean]>=4.9.0
lxml_html_clean>=0.1.0
Pillow>=10.0.0
